# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Goat Nutrition App API
  version: 0.1.0
  description: |
    Minimal OpenAPI for key endpoints. Auth uses cookie session (Flask-Login).
servers:
  - url: http://localhost:5001
paths:
...
  /api/agent/tip:
    get:
      summary: Daily tip (requires X-Api-Key)
      responses:
        '200': { description: OK }
        '401': { description: Missing API key }
  /api/agent/recommendation:
    post:
      summary: Nutrition recommendation (JSON contains api_key)
      responses:
        '200': { description: OK }
        '400': { description: Validation error }
  /api/agent/chat:
    post:
      summary: Chat with AI (text or image multipart; requires api_key)
      responses:
        '200': { description: OK }
        '400': { description: Validation error }
...
  /api/agent/analytics-report:
    post:
      summary: Generate analytics report via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsReportRequest'
      responses:
        '200':
          description: Report generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_html: { type: string }
                  report_markdown: { type: string }
        '400': { description: Validation error }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    AnalyticsReportRequest:
      type: object
      required: [api_key]
      properties:
        api_key: { type: string }
        filters:
          type: object
          additionalProperties: true
        cohort:
          type: array
          items:
            type: object
            additionalProperties: true
        cost_benefit:
          type: object
          additionalProperties: true
        insights:
          type: array
          items: { type: string }
...
security:
  - cookieAuth: []